+++ StereoKitC/systems/platform/uwp.cpp
@@ -46,7 +46,7 @@ using namespace winrt::Windows::Graphics::Display;
 using namespace sk;
 
 void uwp_on_corewindow_character(CoreWindow const &, CharacterReceivedEventArgs const &args) {
-	input_keyboard_inject_char((uint16_t)args.KeyCode());
+	input_keyboard_inject_char((uint32_t)args.KeyCode());
 }
 
 void uwp_on_corewindow_keypress(CoreDispatcher const &, AcceleratorKeyEventArgs const & args) {
